# swagger.yaml
openapi: 3.0.3
info:
  title: E-Commerce API — Testing Practice
  version: 1.0.0
  description: |
    Full-featured e-commerce REST API built for API testing practice.
    Covers auth flows, CRUD, pagination (offset + cursor), caching (ETag/304),
    file uploads, BOLA/IDOR, rate limiting, fault simulation, idempotency,
    RFC 7807 errors, state machines, and role-based access.

    ## Seeded credentials (reset anytime via `POST /v1/seed/reset`)
    | Role  | Email             | Password   |
    |-------|-------------------|------------|
    | admin | admin@test.com    | Admin123!  |
    | user  | user@test.com     | User123!   |
    | user  | user2@test.com    | User123!   |

    ## Notable testing hooks
    - `?_delay=3000` — delays response by N ms (max 10 000)
    - `?_fail=503`   — returns a simulated error with that status code
    - `Idempotency-Key` header on `POST /orders` — prevents duplicate orders on retry
    - `If-None-Match` header on GET product(s) — returns 304 when data unchanged
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://YOUR-APP.onrender.com/v1
    description: Production (update after deploying to Render)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Problem:
      type: object
      description: RFC 7807 Problem Details
      required: [type, title, status, detail, instance]
      properties:
        type:     { type: string, example: "http://localhost:3000/errors/not-found" }
        title:    { type: string, example: "Not Found" }
        status:   { type: integer, example: 404 }
        detail:   { type: string, example: "Product with id 99 does not exist" }
        instance: { type: string, example: "/v1/products/99" }
...
    Category:
      type: object
      properties:
        id:          { type: integer, example: 1 }
        name:        { type: string,  example: "Electronics" }
        description: { type: string,  example: "Gadgets and devices" }
        createdAt:   { type: string,  format: date-time }
    Product:
      type: object
      properties:
        id:          { type: integer, example: 1 }
        name:        { type: string,  example: "Smartphone Pro X" }
        description: { type: string,  example: "Latest flagship smartphone" }
        price:       { type: number,  format: float, example: 999.99 }
        stock:       { type: integer, example: 50 }
        categoryId:  { type: integer, example: 1 }
        imageUrl:    { type: string, nullable: true, example: "/v1/products/1/image" }
        createdAt:   { type: string,  format: date-time }
    CartItem:
      type: object
      properties:
        id:        { type: integer }
        productId: { type: integer }
        quantity:  { type: integer }
        product:   { $ref: '#/components/schemas/Product' }
    Cart:
      type: object
      properties:
        items:     { type: array, items: { $ref: '#/components/schemas/CartItem' } }
        total:     { type: number, example: 249.99 }
        itemCount: { type: integer, example: 1 }
...
    OffsetPagination:
      type: object
      properties:
        type:       { type: string, example: offset }
        page:       { type: integer }
        limit:      { type: integer }
        total:      { type: integer }
        totalPages: { type: integer }
        hasNext:    { type: boolean }
        hasPrev:    { type: boolean }
...
  responses:
    Unauthorized:
      description: Unauthorized — missing, expired, or invalid token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Forbidden — insufficient role
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
paths:
...
  /categories:
    get:
      summary: List categories (paginated)
      tags: [Categories]
      parameters:
        - { name: page,  in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
      responses:
        '200':
          description: Paginated category list
          headers:
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:       { type: array, items: { $ref: '#/components/schemas/Category' } }
                  pagination: { $ref: '#/components/schemas/OffsetPagination' }
    post:
      summary: Create category (admin)
      tags: [Categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string, example: "Sports" }
                description: { type: string, example: "Sports equipment" }
      responses:
        '201':
          description: Category created
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /categories/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      summary: Get category by ID
      tags: [Categories]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Replace category (admin) — full replace, all fields required
      tags: [Categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                description: { type: string }
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/Category' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete category (admin)
      tags: [Categories]
      security:
        - bearerAuth: []
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Products ────────────────────────────────────────────────────────────────
...
  /cart:
    get:
      summary: Get current user's cart
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/Cart' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      summary: Clear entire cart
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        '204': { description: Cart cleared }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /cart/items:
    post:
      summary: Add item to cart
      tags: [Cart]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, quantity]
              properties:
                productId: { type: integer }
                quantity:  { type: integer, minimum: 1 }
      responses:
        '201':
          content: { application/json: { schema: { $ref: '#/components/schemas/Cart' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { description: Insufficient stock, $ref: '#/components/responses/UnprocessableEntity' }
  /cart/items/{itemId}:
    parameters:
      - { name: itemId, in: path, required: true, schema: { type: integer } }
    put:
      summary: Update item quantity
      tags: [Cart]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 1 }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/Cart' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      summary: Remove item from cart
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        '204': { description: Item removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Orders ──────────────────────────────────────────────────────────────────
...